Update README.md
Each USB port is compliant to the usb.org specification. Each USB compliant port must handle 500mA. As Raspberry have two ports, so the total USB power budget is 1A. The previous text can lead the reader to think that the mentioned power requirement is for the whole Raspberry power consumption, not only USB: "The maximum power the Raspberry Pi peripherals can use is 1 Amp. If you need to connect a USB device that will take the power requirements above 1 Amp, then you must connect it to an externally-powered USB hub."
My text suggestion makes it clear that it refers only to the USB peripherals: "The maximum power the Raspberry Pi peripherals can use is 1 Amp. If you need to connect a USB device that will take the USB power requirements above 1 Amp, then you must connect it to an externally-powered USB hub."